refactor: merge effort.ts and model.ts into settings.ts
Consolidate the two single-config-value picker modules into a unified
Settings class that owns runtime-settable preferences (model and effort)
and exposes pickModel()/pickEffort() operations. The command registry
maps /model and /effort to settings methods; Settings has no knowledge
of commands.

Closes #689
